requirements.txt
numpy==1.22.4
rasterio==1.2.10
scikit-learn==1.3.2
tensorflow==2.13.0
matplotlib==3.7.3
pyyaml